Added All Other Buildings
-set up all other building subclasses
-added Dragon class but only health and taking damage so it could be referenced by the Turret

+++ FantasyRebuild/Assets/Scripts/Player.cs
@@ -117,8 +117,8 @@ public class Player : MonoBehaviour
     {
 
     }
-    public void AddProductionBoost(int amount) => totalProductionBoost += amount;
-    public void SubtractProductionBoost(int amount) => totalProductionBoost -= amount;
+    public void AddProductionBoost(float amount) => totalProductionBoost += amount;
+    public void SubtractProductionBoost(float amount) => totalProductionBoost -= amount;
 
     #endregion
 }
